pyproject.toml
[project]
name = "veAgentBench"
version = "0.1.0"
description = "A comprehensive LLM evaluation framework with custom metric support for agent benchmarking"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Bytedance IaaS Team", email = "xiaobao@bytedance.com"},
]
maintainers = [
    {name = "Bytedance IaaS Team", email = "yetan@bytedance.com"},
    {name = "Bytedance IaaS Team", email = "xiaobao@bytedance.com"},
    {name = "Bytedance IaaS Team", email = "liuyanfeng.11@bytedance.com"},
]
keywords = ["llm", "evaluation", "benchmark", "agent", "ai", "custom-metric"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    # Existing project deps
    # "deepeval>=3.5.8",
    "veadk-python==0.2.7",
    "volcengine-python-sdk[ark]>=4.0.3",
    # Merged runtime deps from deepeval (converted from Poetry spec)
    "requests>=2.31.0,<3.0.0",
    "tqdm>=4.66.1,<5.0.0",
    "tabulate>=0.9.0,<1.0.0",
    "sentry-sdk",
    "rich>=13.6.0,<15.0.0",
    "portalocker",
    "openai",
    "aiohttp",
    "typer>=0.9,<1.0.0",
    "click>=8.0.0,<8.3.0",
    "ollama",
    "setuptools",
    "wheel",
    "nest_asyncio",
    "tenacity>=8.0.0,<=10.0.0",
    "opentelemetry-api>=1.24.0,<2.0.0",
    "opentelemetry-sdk>=1.24.0,<2.0.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.24.0,<2.0.0",
    "grpcio>=1.67.1,<2.0.0",
    "anthropic",
    "google-genai>=1.9.0,<2.0.0",
    "posthog>=6.3.0,<7.0.0",
    "pyfiglet",
    "python-dotenv>=1.1.1,<2.0.0",
    "pydantic>=2.11.7,<3.0.0",
    "pydantic-settings>=2.10.1,<3.0.0",
    "jinja2",
    "akshare>=1.17.63",
    "tos>=2.8.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
docs = [
    "sphinx>=6.0.0",
    "sphinx-rtd-theme>=1.2.0",
    "sphinx-autodoc-typehints>=1.23.0",
]
all = [
    "veAgentBench[dev,docs]",
]

[project.urls]
Homepage = "https://github.com/bytedance/veAgentBench"
Documentation = "https://veagentbench.readthedocs.io"
Repository = "https://github.com/bytedance/veAgentBench"
Issues = "https://github.com/bytedance/veAgentBench/issues"
Changelog = "https://github.com/bytedance/veAgentBench/blob/main/CHANGELOG.md"

[project.scripts]
veagentbench = "veagentbench.cli:main"

[build-system]
requires = ["setuptools>=61.0", "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["veagentbench*"]
exclude = ["tests*", "docs*", "examples*"]

[tool.setuptools.package-data]
"veagentbench" = ["py.typed", "*.json", "*.yaml", "*.yml"]

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["veagentbench"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
